LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scope-shadow-names-expected.txt


FAIL anchor-scope matches tree-scoped names assert_equals: expected "10px" but got "1px"
FAIL anchor-scope in shadow does not affect slotted-in element assert_equals: expected "10px" but got "1px"
FAIL anchor-scope in ::slotted() rule does not affect anchoring outside assert_equals: expected "10px" but got "1px"
FAIL anchor-scope in shadow does not affected slotted-in element (:host) assert_equals: expected "10px" but got "1px"
PASS anchor-scope in ::part() affects slotted-in element
